requirements.txt
flask
flask-session
cs50
Werkzeug
Pillow
python-docx
PyPDF2
python-pptx
python-magic
appose-words